ci: optimize workflows (#190)
* ci: optimize workflows with rust-cache and cargo-binstall

* ci: add Code Quality workflow to satisfy branch protection rules

* style: run cargo fmt

* fix: resolve clippy warnings

* fix: import WriteXdr in tests module

* style: cargo fmt after fixing import

* ci: fix quality.yml syntax error and add rust-cache

+++ cargo-budget-report/src/main.rs
@@ -5,7 +5,7 @@ use indicatif::{ProgressBar, ProgressStyle};
 use std::collections::HashMap;
 use std::path::Path;
 use std::process::Command;
-use stellar_xdr::curr::{Limits, ReadXdr, SorobanTransactionData, WriteXdr};
+use stellar_xdr::curr::{Limits, ReadXdr, SorobanTransactionData};
 use tabled::{Table, Tabled};
 use wasmparser::Parser as WasmParser;
 
@@ -35,6 +35,7 @@ struct BudgetToml {
     functions: HashMap<String, FunctionConfig>,
 }
 
+#[allow(dead_code)]
 #[derive(serde::Deserialize, Debug)]
 struct Resources {
     instructions: u64,
@@ -42,6 +43,7 @@ struct Resources {
     write_bytes: u64,
 }
 
+#[allow(dead_code)]
 #[derive(serde::Deserialize, Debug)]
 struct TransactionData {
     #[serde(alias = "resources")]
@@ -334,26 +336,26 @@ fn main() -> Result<()> {
                 continue;
             }
 
-            let (instructions, read_bytes, write_bytes) =
-                extract_metrics(&rpc_resp).context("Failed to extract metrics from RPC response")?;
+            let (instructions, read_bytes, write_bytes) = extract_metrics(&rpc_resp)
+                .context("Failed to extract metrics from RPC response")?;
 
             reports.push(CostReport {
                 package: package.name.to_string(),
                 function: function.clone(),
                 metric: "CPU Instructions",
-                value: instructions as u32,
+                value: instructions,
             });
             reports.push(CostReport {
                 package: package.name.to_string(),
                 function: function.clone(),
                 metric: "Read Bytes",
-                value: read_bytes as u32,
+                value: read_bytes,
             });
             reports.push(CostReport {
                 package: package.name.to_string(),
                 function: function.clone(),
                 metric: "Write Bytes",
-                value: write_bytes as u32,
+                value: write_bytes,
             });
         }
     }
@@ -405,6 +407,7 @@ mod tests {
     use std::fs;
     use std::path::PathBuf;
     use std::time::{SystemTime, UNIX_EPOCH};
+    use stellar_xdr::curr::WriteXdr;
 
     fn unique_test_path() -> PathBuf {
         let mut path = std::env::temp_dir();
@@ -470,8 +473,7 @@ mod tests {
             .join("fixtures")
             .join("simulate_transaction_response_valid.json");
         let fixture_json: serde_json::Value = serde_json::from_str(
-            &std::fs::read_to_string(&fixture_path)
-                .expect("failed to read fixture file"),
+            &std::fs::read_to_string(&fixture_path).expect("failed to read fixture file"),
         )
         .expect("failed to parse fixture JSON");
 
@@ -524,13 +526,15 @@ mod tests {
             .join("fixtures")
             .join("simulate_transaction_response_malformed.json");
         let fixture_json: serde_json::Value = serde_json::from_str(
-            &std::fs::read_to_string(&fixture_path)
-                .expect("failed to read malformed fixture file"),
+            &std::fs::read_to_string(&fixture_path).expect("failed to read malformed fixture file"),
         )
         .expect("failed to parse malformed fixture JSON");
 
         let result = extract_metrics(&fixture_json);
-        assert!(result.is_err(), "extraction should fail on malformed response");
+        assert!(
+            result.is_err(),
+            "extraction should fail on malformed response"
+        );
     }
 
     #[test]
@@ -543,7 +547,10 @@ mod tests {
             }
         });
         let result = extract_metrics(&rpc_json);
-        assert!(result.is_err(), "extraction should fail when transactionData is not a string");
+        assert!(
+            result.is_err(),
+            "extraction should fail when transactionData is not a string"
+        );
     }
 
     // --- Budget toml loading tests ---
